[debugger] Index in anonymous namespaces.
Previously we would skip all namespaces without names and would
therefore omit anything contained in them from the index.

This adds support for indexing items in anonymous namespaces as if the
namespace wasn't there. This handles most cases the way users expect. A
comment was added about how this might be changed in the future.

Added a test. The checked in DWARF binary was not updated (I don't
expect to keep this in sync with the code). A test using this depended
on a constant that changed, so this test was modified to avoid using the
constant so the code and checked-in binary can diverge.

Change-Id: If3f45e6a4a6ac997306c289f22c0e322d0c810af