[fidl] Naming naming scheme: with dot slash it is 'flat name'
In an effort to standardize how names are externalized by fidlc, renaming
and consolidating name generation.

Use of names with dot slash as separators is currently used for normal
flat name output. This is used in four places:

1. Error reporting
2. JSON IR
3. Coded tables (for runtime debug)
4. Ordinal calculation for methods

We want to move to slash dot (instead of dot slash) for all, and plan to
add a NameFlatNameNew to transition. Changing (1) is easy, changing (4)
will be done as part of ordinal increase from 32 to 64 bits, changing
(2) requires touching the various backends, and changing (3) is unknown
today, assumed to be easy.

Test: no behavior change

Change-Id: I69cc2d93032226b7ea2d23931585c61acc8ee0b1