[kernel][object] Fix interrupt dispatcher test
This fixes an underlying functional issue in the test where the created
InterruptEventDispatcher was not being correctly shut down due to
moving the dispatcher out of the KernelHandle, causing on_zero_handles
to not get called. The underlying dispatcher is actually not needed so
we simply avoid releasing it from the KernelHandle.

Also fixes some stylistic issues of the thread signaling with an
explicit shared atomic instead of abusing thread_kill.

ZX-4305 #done

Change-Id: I104f322c5365fdcad5d2ee556c0536f6c1ba98f4