[debugger] Add "run" documentation.
Adds a warning when using "run" that it may not work. We see more
processes that don't work and people are confused.

Adds significant documentation in "help run" about the issue. This is
mostly copied from the markdown documentation.

Adds documentation on what exactly filters match.

Fixes a error in component launching that adds the component URL. This
is not done using the normal component launch flow.

DX-1603

Change-Id: I2cbaafcc39fcf99cc854fd4e329e53db38500c59