[testing] Make qemu.go flush after sending RunCommand
Without a stdin Flush after sending a command, it's not possible to
interleave a set of RunCommand()s and WaitForLogMessage()s as the input
is arbitrarily buffered before it reaches qemu. With the flush, we can
write a more "expect" style of test as the interface implies.

ZX-3637 #comment [testing] Make qemu.go flush after sending RunCommand

Change-Id: I7275ef021276dbcc1ea1e68c39c7d18129e027e7