[zx] Split mexec functionality out from devmgr.
Introduce new fuchsia.kernel.MexecBroker that will live in svchost,
similar to f.k.DebugBroker. It will take over responsibility for
providing the service that is used to mexec the system.

This also changes slighly how the Suspend call works for devmgr. The
suspend call will now wait until the suspend operation is done before
returning a result to the fidl caller. The mexec process now uses this
functionality to allow a caller to call the mexec syscall after all
devices are suspended instead of relying on a special case in devmgr to
do it.

Issue: ZX-3403 #Mexec on its own.
Test:
 * Added unit tests
 * paved vim2 and astro twice (to ensure zedboot was updated).
 * mkzedboot and pave eve.

Change-Id: I84fb42a70de039b32b71db08242701d2a66e1c53