[mdns] fix publication ordering issue
This CL fixes a problem in which a publishing 'agent' attempts to
send records before it has been started. The agent is supplied with
a set of configured addresses on start. Without these addresses, the
attempt to send records failed.

A DCHECK was in place to detect attempts to access the address config
prior to start, but it wasn't working, because the pointer to the
address object wasn't set to null initially. This has been fixed, so
a regression in behavior will be detected.

TEST COVERAGE: This CL introduces a netemul test that verfies that
mdns can talk to another instance of itself. The test addresses
publication, discovery and host name resolution. While this test
doesn't verify that the implementation is compliant with mDNS
standards, it is sufficient to detect a regression of the bug fixed
by this CL.

Change-Id: I58b78248e8c55555cf6d301d875f55a05af38730