[musl][stdio] Fix racy crash in FILE* access
Multithreaded access to FILE* structures was not correctly locked when a
thread involved with locking had the high bit of its handle value set.

As handles reserve the low two bits, create the identifier for the lock
by shifting the handle value down so that the sign bit will never be
set.

ZX-4278 #done

Change-Id: Ib30f66f86dd30a03926443f477b694773922a590