[build] Add zircon_extra_ninja_targets build argument
This lets the controlling Fuchsia GN build specify additional Ninja
targets that should be in the default set for the subordinate Zircon
build.  This is needed for some complex non-default build scenarios
that require fresh kludges before we achieve the unified GN build.

Also moved out of zircon.gni the remaining pieces that are only needed
in //BUILD.gn and nowhere else.

Bug: BLD-514 #done
Change-Id: Ic363025bf813c9a35e198192576689f907ffebb4