[graphics][compute]: Generate shaders for debugging.
This CL refactors hotsort_target.gni and spinel_target.gni
to use a common GN template to compile compute shaders to
the SPIR-V format, as well as allow for the generation
of human-readable versions of the generated shader sources.

A new declare_args variable that can be set to true in
your out/default/args.gn file to force the generation of
pre-processed and auto-formatted compute shaders for hotsort
and spinel.

These are placed under the debug/ sub-directory, as in:

  out/default/gen/src/graphics/lib/compute/
     hotsort/targets/hs_amd_gcn3_u64/
       comp/
         hs_transpose.comp    -> original shader
       debug/
         hs_transpose.comp    -> same, but much easier to read.

This helps understand the impact of hotsort_gen parameters
and vendor-specific configurations, or simply what hotsort
and spinel modifications end up doing to the generated code.

These shaders are never used, and are only for manual review,
though it would be technically possible to compile them.

+ Move some build settings to gn/build_settings.gni to prepare
  for SPN-35.

TEST: Rebuild with fx build src/graphics/lib/compute:tests

Change-Id: Iad1c3b408e5f4667facd5058ecf5084d48a43da8