[componentmgr] Implement out directory in Hub
This CL makes a number of changes to support the out directory in the Hub.

1. It clones the client-side DirectoryProxy of the outgoing directory held
by component manager to use in Hub.

2. If the clone is successful (it isn't always available in tests), then
it adds an out directory in the exec directory via a DirectoryBroker. DirectoryBroker
did not support following paths into another directory tree, and so it was updated
to support passing along a relative path.

3. All calls to Controller::add_entry in Hub were replaced by Controller::add_entry_res
to avoid races in tests.

4. A unit test was added to verify that an out directory can now be hosted in hub.

Bug: CF-541

Change-Id: I41550d6eb2d59cd4b04f6a31ba44c097ec173650