[debugger] Add CFA to StackFrame
Make the unwinder compute the "canonical frame address" and add it to
the StackFrame structure. This is equivalent to the "frame fingerprint"
which was computed on the client side without having complex code to
handle asynchronous computation of it. This substantially simplifies
code in the Stack object.

Expose this CanonicalFrameAddress in the Frame object.

Hooks up the CFA in the DWARF expression evaluator. This will now
evaluate most frame base expressions from GCC and allow many more
variables to be printed in GCC-compiled code.

This change required a lot of test updates that send stack information.

Updates error messaging for some other common GCC dwarf expressions
opcodes that we don't support.

Adds the frame base pointer to the verbose frame format.

Change-Id: Iac981a6b829e2c6b357ebae8f1d7dfd5c8284785