[wlan][kde] Add tests for KDEs and provide Writer
This is a re-landing of: I3372ddbe8cf62486a806505bfb7276469e37dc1b
The CL was reverted due to connection issues with some TP-Link
routers. Revert in Ib4643265dc5c39a87970780780ef7b5de5fef039
The reason for the routers in question to fail was that the CL
enforced all key data to be padded to meet IEEE specifications.
However, the key data is only required to get padded if it'll later
be encrypted. Because the second message of the 4-Way Handshake is
not encrypted the padding wasn't required and TP-Link failed parsing
the frame.

Test:
Updated existing tests to catch such a regression in the future.
Added a test for the the newly added plaintext function.

-- Original commit message ---
Provides a convenience Writer type for constructing key data.
This writer enforces that necessary padding is always part
of the written key data.
Adds tests for verifying key data is written properly.

This is part of test hackathon to improve RSN testability.
WLAN-1132 #comment

Test=Included

Change-Id: I7cea02911393a24c4881e6939758754d85e86ce7