[netstack3] Skeleton to handle IPv6 Extension Headers
BUG: NET-2240 #progress

Created the skeleton to handle IPv6 extension headers. This change does
not introduce any new functionality. It simply includes stubs that can
be used to handle extension headers. However, unlike before, we will now
panic when attempting to handle an extension header we have no
implementation for but successfully parsed, because normally, if we do
not fully support some extension header or option, we should return the
appropriate parameter problem code.

Test: Existing tests continue to test the handling of extension headers
we already support parsing of AND know how to handle (hop-by-hop options
and destination options). For other extension headers (fragment and
routing), the stubs panic as they don't have an implementation yet.

Change-Id: I09e6fd4bed7587dea0b33aa6e7933912d39f9e3a