[pi][test] Add a small app exercise sync obj PI
Add a small app which sets up various scenarios in which priority
inheritance should come into play.  Add tests for chains of mutexes,
multiple simultanious mutex waiters, and multiple condvar waiters who
are woken in a broadcast fashion.

Add template adapters for...
++ libsync mutexes and condvars
++ pthread mutexes and condvars (with and without PI)
++ C11 mtx_t and cnd_t
++ fbl::Mutex

Note, this is not an automateable test.  One needs to run the test and
use the generated kernel trace events to verify that PI is occuring as
you would expect.

Change-Id: Iee5a278983523c83385e8c3f707e40bc9d43724c