[dhcpd] Add and Use SubnetMask struct
SubnetMask abstracts over the two common ways of representing a
Subnet Mask: as an integer less than 32, and as an array of four octets.
This abstraction addresses NET-2321, which is caused by supplying an integer
when octets are expected. SubnetMask is now used when supplying a Subnet
Mask DHCP option to a DHCP client. The DHCP configuration file and
integration test have been updated to use a 25 bit mask in order to
verify that the client can receive a non-default mask from the server
and still behave normally.

NET-2321 #done

Change-Id: I8120b6d4ae431dcc3432812403b45ab70c0b0b4b