[vulkan] Fix VK_FUCHSIA_external_memory usage
DX-939 #progress
MA-394 #progress

VkExternalMemoryImageCreateInfo should be specified when
creating images that will be backed by memory that can
be exported or imported.

No change in behavior. Just following the spec more accurately,
which is important for future hardware support.

This will allow us to remove workarounds in the goldfish
vulkan ICD that are currently required for nvidia hardware.

Test: fx shell run fuchsia-pkg://fuchsia.com/vkcube_on_fb#meta/vkcube_on_fb.cmx
Test: fx shell run fuchsia-pkg://fuchsia.com/vkcube_on_scenic#meta/vkcube_on_scenic.cmx
Change-Id: I4c203607470991f18c966bfe04ad95becf75b2fd