[tracing][escher] use vkTimestamps correctly in TimestampProfiler
We no longer assume that the timebase given to us in the VkTimestamp
is the same as the CPU's. While the previous version worked on available
hardware, it led to undefined behavior on AEMU which made the GPU events
on the traces unusable.

Instead, we now use the fact that CleanupEscher() runs every millisecond
to offset the durations we get back from the GPU by 0.5ms, guaranteeing
us to be precise to +/- 0.5ms.

Ideally we would use a Vulkan timings extension to get even more precise
data, but this should be fine until then to get GPU work included in AEMU
flows. See SCN-1460 for details.

SCN-1460 #done

Test: unit tests, tracing on real hardware and AEMU

Change-Id: I3766ceebca686c5a2f20746afee78f4f7c5a29db