[mtk][usb] Rename Endpoint type to RequestQueue
This renames the Endpoint class to RequestQueue to better describe its
actual function.  I was working on adding hub support, and the diff
generated by renaming usb-endpoint.h/cpp as well as writing a new
usb-endpoint.h/cpp was ludicrous.  This performs a standalone
refactoring.

This is a 0-sum patch with no change in behavior.

Test: runtests -t ...
Change-Id: Ic9d6004a9da6b81b2c25df4906179b6b31fec228