[netemul] Fix flaky test SandboxTest.SyslogWithNoKlog
BUG: FLK-350 #done
BUG: FLK-355 #duplicate
BUG: FLK-356 #duplicate

Fix a race condition with netemul's SandboxTest.SyslogWithNoKlog test,
which caused a flake.

The test waits until it receives a log message (without receiving any
kernel log messages) then publishes an event to the bus. The DummyProc
sends a log message then connects to the bus and attempts to wait for
the event. However, after the log message is received by the test, the
test could publish the event that DummyProc would wait on before it
connects to the bus and starts to wait for the event. This change
connects DummyProc to the bus right when it starts, before doing
anything of interest so that it doesn't miss any events.

Race:
- DummyProc sends a log
- Test receives log
- Test publishes event=100 on the bus
- DummyProc connects to the bus
- DummyProc waits for event=100 (but the event has already been
  published)
- Test waits for DummyProc to terminate (but it never does because it is
  waiting for event=100)

Change-Id: I67e20a96a67f0bef24bb84197df5f318a8351b17