[fidl][lint] CLI options, include/exclude checks, order findings
Integrated the Fuchsia cmdline::ArgsParser and added command line
options to support near term use cases.

Since some checks are "noisy", added options to exclude certain
checks, or include only a certain check(s), depending on the type
of analysis needed.

Findings are not always added in source position order, so Linter
now adds Findings to a multiset, with a comparator, and later
transfers the Findings in the correct order to the caller's
Findings out parameter.

Test: fidl-compiler-test
Change-Id: I52ac7eeb9ebffcf8dba62fb0978fcd055082a368