[astro][sherlock][usb] OTG support for Astro and Sherlock
The aml-usb-phy-v2 driver now supports switching between host and peripheral role
automatically based on what is connected to the port.

The new dwc2 driver is used for peripheral role support.

Astro and Sherlock board drivers configure the peripheral stack to expose
a composite interface containing both CDC ethernet and support for usb-peripheral-test.

Added a new banjo protocol for the dwc2 driver to inform the PHY of the current connection state.
This is needed to ensure USB stability on Astro and Sherlock.

TEST: usb-peripheral-test succeeds for both Astro and Sherlock.
      Paving is stable for Astro and Sherlock, both in host and peripheral roles.
Change-Id: I45e570ce43a0e64fb93c195f863be5e16446cae3