[zircon][arm64] Watchpoint kernel support.
This adds the capability to set up watchpoints from the write_state
syscall and have it correctly install within the thread in question.

This logic mimics very closely what the HW breakpoint code does.

This also adds a handler for the HW exception in the arch specific code.

Added a minimal repro case that sets up a HW watchpoint and verifies
that it's hit.

Change-Id: Id0300d37fa7909f48bf71f82c882d9075d4a5293