[amber] Recover from partially installed packages
This change fixes an issue in Amber where if a package's meta FAR exists
but is not indexed, package installation would fail while attempting to
write the meta FAR (which it can't do because it already exists).
Instead, after writing the meta FAR (and informing pkgfs about it),
Amber will now continue the installation if pkgfs presents a needs
directory for the package or complete installation if the package
becomes readable.

Test: amber_install_test

PKG-772 #done

Change-Id: I795b0e930b88ffa1fd11eda22212bb3705c92d70