[pave-archive] add embedded bootserver arguments
As described in the comment above the argument declaration, this enables a
very specific fudge which is used in some board definitions to workaround
issues between poorly designed hardware mixing with software bugs that leads
to paving failures. A related root cause investigation for the particular
fudge is detailed at length in CONN-50. The workaround is being moved into
board configurations so that it can be applied to more devices while putting
the configuration in a more sensible location.

New or additional uses of this build flag should be avoided without further
discussion due to potential brittleness.

Bug: ZX-4283 #comment export declared argument for bootserver args
Change-Id: I3985bd802800c74662ab371e2139bd681f1986de