[libsync][pi] Implement priority inheritance in libsync.
Implement priority inheritance behavior in libsync, both in mutexes as
well as the condvar/mutex interactions.

Note: condvar PI is still more or less incorrect.  The current
implementation maintains its own wait queue in user-land memory
instead of using a futex for this purpose.  As a result, in a
broadcast situation, waiters are let out of the user mode queue in the
order that they joined, and are transfered (one at a time, via
futex_requeue) to be waiters on the mutex.  As a result, the holder of
the mutex only feels the pressure of the most recently transferred
waiter, not all of the waiters.

The fix for this would be to eliminate the user mode wait queue and
just use a futex for this purpose.  This, however, would change the
wakeup ordering in single-signal use cases.  If the order of wakeup
*must* match the order of queueing, then a futex would not do the job
as it reserves the right to decide the order of release when
individual threads are woken.

Tests: mutex_pi_exerciser and validation using kernel tracing
ZX-1798  #comment libsync mutexes converted

Change-Id: Ibc04f12e827c829970a116a53246663523c23464