[kernel][vm] Simplify clone observability behavior
If a parent vmo is shrunk and then reenlarged, don't expose the new
pages to the childen, even if they were originally visible to those
children. This gives simpler semantics for which parent pages are
visible to a child, as it no longer depends on what pages were written
to during the period during which the parent was smaller.

Change-Id: Ia855a07c5176931f57cd9c19e50c41fb2c3e3141