[appmgr] Fix use after free in system_objects.
There are multiple callbacks capturing "this" in system_objects, which
causes use-after-free errors. A new test case is included that fails
consistently with --variant=asan/appmgr.

This change shares the one variable that needed this (process_) as a
shared_ptr to all closures. This is a temporary standin until CF-761 is
closed.

DNO-614: #done
Change-Id: Ie20553a5fab8d99475e78840d6d88dca9b770f6c