[fvm][client] Add BlockDevice-based interface to ResetAllSlices
Fuchsia filesystems are transitioning how they refer to their underlying
block devices.

- Today: A combination of file descriptor and channel. "ResetAllSlices"
has an interface which reflects this usage.
- Soon: Exclusively using the BlockDevice interface. This will provide
easier mechanisms for mocking and testing the filesystem's usage of the
device.

This patch adjusts "ResetAllSlices" to a purely channel-based interface,
and provides "ResetAllSlices2", which operates on a BlockDevice
interface. Once all filesystems refer to their underlying device via
this interface, the old function will be removed.

TEST: Tests for ResetAllSlices added.

Change-Id: I088f3fd76097f3daac379cc534439c075a5ebd89