[aml-sd-emmc] Wait for the interrupt thread only once
Reland of https://fuchsia-review.googlesource.com/c/fuchsia/+/289168
with the debug log removed.
Destroy the interrupt and wait for the interrupt thread in the
destructor. We anyway call the desctructor in release.

Test: iochk; wlan scan,connect,ping.
I added a lifecycle test for aml-sd-emmc. runtests -t aml-sd-emmc-test.
More holistic ddk compatibility tests are on way

Change-Id: Ic276c5d5e7b47af8dfe4d8d4cba86387c6548df9