[x86] Support entering deeper C-states with MWAIT.
Allows a per-microarchitecture configuration to be defined which lists
all C-states supported by the microarchitecture, as well as some
metadata for each state to make intelligent decisions about which state
to enter:
- exit latency (microseconds to exit the state)
- can flush TLB

Currently the policy is to always pick C1, which makes this change a
NOP. A later change will add intelligent state selection.

Adds a list of C-states for Kaby Lake, and stubs out all other
microarchitectures. This list was derived from:
- The intel datasheet for Kaby Lake (for state list and TLB flushes)
- ACPI _CST entries (for exit latencies per state).

Note that the C-states in this change are processor C-states, not ACPI
C-states.

The Kaby Lake exit latencies come from the ACPI tables. The mapping from
processor C-states to ACPI C-states is defined in the Intel datasheet.

Also adds the `k idlestats` command to report on which C-states were
entered in the idle loop, and the mean duration the system spent in each
state while idle.

ZX-2326

Change-Id: I15a834c2966cc7ed3ccc6cdcec90be80d4d9c901