[debugger] Start on new formatting system.
This is a first iteration on a new formatting system that will allow
more flexible console formatting (line wrapping, etc.) as well as
better support future GUI integration and pretty printers.

This code adds the basics of the FormatNode class. This class contains a
detailed description of the design in the header file.

Adds numeric formatting code. This copies the current formatting code
and unit tests from the old formatter. This current code is not yet
hooked up to anything. In a followup the old formatter will call into
the new one for numeric types. We will then move formatting for types
over to the new system one-at-a-time until the old system is a shell
that can be deleted.

Change-Id: I2ff9503706c36578301e1dc96dcdecda33023985