[musl][pi] Implement priority inheritance in MUSL.
Implement priority inheritance behavior in MUSL for both C11 and
pthreads, and for both both mutexes and condvar/mutex interactions.

Note: condvar PI is still more or less incorrect.  The current
implementation (both in pthreads and C11) maintains its own wait queue
in user-land memory instead of using a futex for this purpose.  As a
result, in a broadcast situation, waiters are let out of the user mode
queue in the order that they joined, and are transfered (one at a
time, via futex_requeue) to be waiters on the mutex.  As a result, the
holder of the mutex only feels the pressure of the most recently
transferred waiter, not all of the waiters.

The fix for this would be to eliminate the user mode wait queue and
just use a futex for this purpose.  This, however, would change the
wakeup ordering in single-signal use cases.  If the order of wakeup
*must* match the order of queueing, then a futex would not do the job
as it reserves the right to decide the order of release when
individual threads are woken.

Tests: mutex_pi_exerciser and manual validation using kernel tracing
ZX-1798  #comment MUSL mutexes converted

Change-Id: Ib49529aa3e9112dee6dea6a06e2f1831a2ef651b