[pseudo-fs] Fix `poll` execution order for directories
Our unit tests for directories did not check for termination of the
tests.  As a consequence a few tests where actually blocked and we did
not notice that.

They were blocked because of an error in the execution order in the
`poll` method for `directory::Simple` (but possibly in other directory
implementations as well).  After we add new connections to the directory
node we did not call `poll_next` on the connection list, and returned
`Poll::Pending`.  This casued the future to be considered as one that
can not make any more progress while in fact it could.

This change fixes the bug, makes sure that directory tests do check for
proper termination and fixes a few tests that did not complete due to
some minor issues.

Change-Id: I8ed9aac89813942ff9ff1e2bb8e361a0f207c195