[build] Re-enable unused-lambda-capture and user-defined-warnings
These were disabled in the past because our codebase wasn't clean
with respect to these warnings, but that's no longer the case.
Re-enable these warnings to make sure we don't regress again in
the future (note that these warnings are enabled automatically
under -Wall so there's no need to enable them explicitly).

Bug: TO-99
Bug: TO-100
Change-Id: I4754db0b819ada4b85476d0db3809631ca2d2670