[build][fuzzing] Fix fuzzing unit tests
This CL fixes the fuzzing unit tests introduced in CL 279988. With that
CL, the unfortunate choice of a the variable 'package_name' meant the
'generated_test_package' field was ignored and the tests were put a
package with a name that conflicted with the fuzzers package. This
passed CQ because it by default doesn't build the variant fuzzers
package.

Bug: SEC-224
Test: 1) Things should build both with or without `--fuzz-with asan`:
    fx set core.x64 --with //bundles:tests [--fuzz-with asan]
    fx build
  2) For each $fuzzers_package, there should be a test package:
    fx run-tests <fuzzers_package_tests>
  3) For each fuzzer in a package with $fuzz_host == true, there should
     be a host fuzzer test:
     fx run-host-tests <fuzzer_test>
Change-Id: I8bf81f029dfcff9ed6465d81199b56e4b32a2533