[async-testing] Polish up API prior to adding to the SDK
This CL explicitly deletes several copy and move constructors and marks
a number of classes as final.  These changes are easier to make now
before we add this library to the SDK.

Eventually, we should write up a C++ API rubric that covers these
patterns.

Also, remove the * from the owners file to tighten up the governance for
this library.

Test: No behavior change.

Change-Id: Iba9f3ae80cdaad8ee3a1df68698acd04deaa243f