[netstack3] Get IP packet header bytes without a builder/serializer
Get the header bytes for an IP packet without needing to create a
builder or serializer. This is useful for when this netstack does not
yet know how to properly parse or serialize some options or extension
headers but we need to keep them anyways when reassembling a fragmented
packet.

This change also makes the `builder` functions for IP packet types
available only in a test environment as they are not meant to be used in
production.

Change-Id: Id627f58eefa699d49309c201a01363b612c652cc