[debugger] Index main functions.
Searches for functions marked by DWARF as the main function and keep
them in a separate place in the index that can be queried.

Add support for querying the main function with the string "@main". This
is defined to use the marked function if there is one, and fall back to
a global function named "main" if there isn't.

Change-Id: I1b0aa29be45d3ff68422bb610651146e27d79b70