[fidl][lint] Address unexpected case translation
The linter reported that a constant "PPPoE" should be changed to
UPPER_SNAKE_CASE, and suggested "PP_PO_E".

A test case has been added for this, with a comment explaining
that the translation is consistent with other more intuitive
translations.

I tried changing the algorithm, but it made many other translations much
worse. (For example "URLLoader" should be "URL_LOADER", not
"U_R_L_LOADER".)

We can continue to debate and update the translation algorith as needed.
I just wanted to make sure we captured the case and current behavior in
the test.

Test: utils_test
Change-Id: Id26c1ff8b056af39d462a82f5fce3aea6b3490c0