[media][tests] AudioLoopbackTest: Wait for device add in setup
Waiting on the virtual_audio_service to tell us that it has
created the virtual interface is not a good way to tell whether
audio_core has enumerated the new device and begun using it.

Instead, register with the device enumerator and wait for it to
notify us that it has found the device and made it the default.

This test will still be flaky on any product configuration that
launches an audio client at boot, as both the host instance of
audio_core and the componentized version running in the test
environment will still be sharing the dev manager instance and
will race to grab the virtual audio device created.  This is
tracked in MTWN-331.

Test: Ran test in infinite loop on 2 virtual and 2 physical
      machines overnight.  Prior to this patch, it would
      generally fail within the first 1000 iterations on
      any device.
Change-Id: Iac11a18df5f0508982633639c26af4064265a887