[scudo][mutex] Implement priority inheritance for scudo mutexes.
More specifically, change the BlockingMutex implementation in scudo so
that it uses sync_mutex_t instead of implementing it's own algorithm
and using futexes.  sync_mutex_t in Fuchsia currently expresses PI, so
now the scudo locks do too.

Also note, while this appears to bring in a dependency on libsync,
MUSL already depends on libsync (in fact, mtx_t is a libsync mutex
under the hood), so in reality, there are no changes to the
dependencies.

If this starts to be any sort of performance problem, I can come back
here and rework this to have a custom implementation which implements
PI by encoding the mutex owner thread handle in the futex storage.

Tests: runtests

Change-Id: I945a00869ee4b38ae123857fafe72d4ad9c6e564