[graphics][compute]: Fix host tests build.
A missing dependency in the action used to generate debug shaders
meant that clean builds could fail if the corresponding action
would run before the generation of the host tool that it uses.

This would result in random build breaks, which would eventually
disappear once the rest of the build caught up and generated the
corresponding host tool.

This CL fixes this very annoying issue!

TEST: fx clean-build src/graphics/lib/compute:tests
Change-Id: I50e0b9afb1e99235d3c5165f6e11fbbc02e0e475