[garnet][rust][packet] Properly parse mutable packets from a mutable buffer
Previously, when attempting to parse a mutable buffer into a mutable packet,
the immutable buffer parsing methods were used. This change fixes that so a
mutable buffer is parsed into a mutable packet by following the mutable
buffer parsing methods.

Test: Updated test mutable buffer parsing methods to modify the buffer and
verify that the original buffer was modified after calling the mutable
parse methods.

Change-Id: I2b85db36248c8d5a38fe4d1bf13246ae52e95637