[astro-display] Use a better timestamp for vsync
Kernel itself tracks the signaled timestamp of the irq
so we can use that instead of the time when the vysnc thread
woke up.

This brings this code in line with mt8167s-display.cpp
as well.

TEST = ota astro device, graphics work. This test does
not change behavior, just the precision of the timestamp
for a system under significant load.

Change-Id: Ie263404e75acee810b3f8533d770ff954ee1260d