[kernel][vmar] Add flag to allow mappings to fault
This introduces a flag for zx_vmar_map that will replace the require
non-resizable flag and make it so that the default behavior of mappings
is to not be able to generate faults.

Although the flag is currently a noop, this change updates the
documentation to be consistent with the eventual behavior. It also adds
the flag to a handful of locations which out-of-tree clients depend on.

Change-Id: Ie61342f54c950271e970256be0c4b5c7007fdb13