[debugger] Print Rust variants.
Read variant and variant_part entries from DWARF.

Make the discriminant optional on the Variant object. This encodes the
default variant which Rust uses for some types. Update the VariantPart
object to pass the parameters in the constructor which makes the
interface a little cleaner. Also add DataMember constructors.

Adds a resolve_variant file which will convert a VariantPart and an
ExprValue it's in to the active Variant. Add tests.

Adds a Rust enum formatter for the new FormatNode-based system. Adds
tests.

Adds a formatter to the current formatting system. This just gets the
output from the new system and reports the resulting string. This will
get deleted when the formatting systems are switched over.

Change-Id: Icba5f9cb1198385f7bbd8f7511843c401029533a