[debugger] Add Variant[Part] symbol objects.
These objects correspond to the DWARF tags DW_TAG_variant and
DW_TAG_variant_part.

Currently this is not hooked up to anything so there is no behavior
change and no tests. This only adds the data model classes. It will be
hooked up and tests added in a second pass.

Change-Id: I8e8a3aa96f71dfc88d3edb2c5431ff845395c760