[banjo][syzkaller] Handle basic syscalls
This CL adds the capability to generate basic sycalls with builtin types
or types from zx.banjo. These types are translated to the types supported
by the Syzkaller [0].

For input arguments, comma-separated argument names with types are
outputted in Syzkaller-supported format [0]. For output arguments,
comma-separated types are outputted, for now.

[0]: https://github.com/google/syzkaller/blob/master/docs/syscall_descriptions.md

SEC-321 #done

Test: fx run-host-tests banjo_unittests

Change-Id: I443f6789f869818549984c13bda6272e84a92db3