[netdump] Netdump packet filter nodes
This is CL number 1 in a series that adds packet filtering capability
to netdump using a filter language with similar syntax to the Berkeley
Packet Filter (BPF).

The series begins with defining filter nodes that match on particular
fields in packet headers. Each node is constructed from a simple
specification that describes the field and value to be matched.
When a filter node performs a matching operation on a packet, a
verdict is returned on whether the packet should be accepted.
Nodes can be joined together with logical operations to form
filter trees that specify more complex match conditions.

Bug: CONN-99
Test: Build with `--with-base //garnet/packages/tests:zircon` as a
parameter to `fx set`, then `runtests -t netdump-test` on the target.

Change-Id: I3fc317d2ef48eeb224eaea0f59b1eb23af7ee34a