[debugger] Step over more code with no line info.
The DWARF line table can have entries that indicate "line 0". These do
not correspond to any source location but rather compiler-inserted
bookkeeping.

This happens more with Rust which seems to generate more line entries
with a "0" line number.

Specifically this change fixes the case where a function call that is
being stepped over returns to an instruction marked with line 0.
Previously it would stop at that location (because the line doesn't
match the one being stepped) and show assembly. With this patch it will
step over the instruction. This is the normal behavior for "line 0"
instructions.

Change-Id: Ifdf08f6795c9139ac9d4ca2d3480c30b0e350c4a