[debugger] Added client state to debugged thread.
Now the debugger tracks what the client wants each thread to do. This is
because correct stepping will stop and resume threads in order to
correctly step over.

This means that the debugger needs to know which threads should remain
suspended when starting them again.

Removed static variables from a test. This would make it fail when run
several times in a row with `--gtest_repeat`.

TEST=No functional change. All debugger tests passes.

Change-Id: Ib8e5111b3e79a9fbeb0b877f161f8d052e1d92ab