[netdump] Netdump packet filter tokens
This is CL number 2 in a series that adds packet filtering capability
to netdump using a filter language with similar syntax to the Berkeley
Packet Filter (BPF).

This CL implements the lexer for input filter strings.

The first step to understanding a filter string input by the user is to
convert it into a sequence of tokens. The lexer in this CL identifies
keywords and literals in the input string, minimizing the need for a
parser of the filter language to handle raw strings. This allows
better efficiency during parsing, for example decisions are made based
on token rather than string comparison. Also the syntax of the
language is more flexible as the lexer forms the central configurable
repository of what filter keywords or named ports are recognized.

The following keywords are supported by the lexer as the starting set
in this CL. It should be straightforward to add further keywords.

  Operations: `(`, `)`, `not`, `!`, `and`, `&&`, `or`, `^^`,
              `greater`, `less`.

      Fields: `ether`, `proto`, `host`, `src`, `dst`, `port`,
              `portrange`.

L2 protocols: `arp`, `vlan`.

L3 protocols: `ip`, `ip4`, `ip6`, `icmp`.

L4 protocols: `tcp`, `udp`.

 Named ports: `dbglog`, `dbgack`, `dhcp`, `dns`, `echo`, `ftpxfer`,
              `ftpctl`, `http`, `https`, `irc`, `ntp`, `sftp`, `ssh`,
              `telnet`, `tftp`.

Bug: CONN-99
Test: Build with `--with-base //garnet/packages/tests:zircon` as a
parameter to `fx set`, then `runtests -t netdump-test` on the target.

Change-Id: I74201dc3513c5a1bc821fd867660b86a4c625800