[intel-hda] Increase timeout waiting for hardware reset.
Increase the timeout waiting for Intel HDA to complete hardware reset
and ring buffer reset from 10µs to 2s.

Repeatedly starting and stopping of the Intel HDA controller showed that
10µs was short enough to occasionally cause false timeouts (~1 in 1000
times), presumably due to thread scheduling, and not the hardware
itself. In turn, this would preventing the driver from coming up
successfully, meaning the device would have no audio.

This CL substantially increases the timeout to 2s. The high value
shouldn't ever be reached in normal circumstances, but only when the
hardware is failing to function correctly.

Change-Id: I1e508920c2c80964e6f096740af9f022da56c23f