[debugger] Add language information.
Adds a way to get at the compilation unit for a given symbol. This
stores some information including the language that the file was written
in.

Add a symbol object model class CompileUnit to represent this. Adds
decoding for this object and a test that the language and name is
decoded propery.

Adds a language bitfield to the expression tokenizer and marks the
existing tokens as applying to Rust or C-like languages. A test is added
for this but nothing is done with Rust as far as expression evaluation
yet.

Adds support for the basic builtin rust types.

Change-Id: Ie1d711b6dc60106c26a4e0614c39e98438eac38a