[ledger] Allow a LedgerRepository to be closed
Adds |Close()| to the LedgerRepository interface, allowing a
LedgerRepository to be shut down cleanly.

The implementation I suggest here waits for each page and ledger to be
empty/idle before closing the connections, and does not close Ledger and
Pages proactively. This ensures that all the user data is in before we
close; otherwise, there is always a risk of data loss. However, we also
wait for the synchronization to be idle, which has advantages (data is
synced) but also drawbacks (we don't have any guarantee this process
will finish quickly).

LE-745 #done

Change-Id: I59dbe4cab8de0fd306402f3aaaa8ebab6efe44cd