[kernel][vmar] Add flag to allow mappings to fault
This introduces a flag for zx_vmar_map that will replace the require
non-resizable flag and make it so that the default behavior of mappings
is to not be able to generate faults.

Although the flag is currently a noop, this change updates the
documentation to be consistent with the eventual behavior.

This relands the kernel portion of ee5c557. That change failed because
some downstream tests currently run newer packages on an older system,
and the older kernel rejects the args passed in by the newer libc.

Change-Id: I4b04ce4af975526e422ed68d9037075c72797e1b