[component_manager] Show errors in routing_integration_test
Previous version would actually never timeout, as 'read' is a blocking
call.  And it will only stop for an explicit match.  Any unexpected
output will cause it to just wait, even when it can not possibly match.
The unexpected output was not shown.

The timeout code makes it look like the test may timeout while in fact
it is very unlikely, as the most likely bocking operation is the `read`
call.  Besides, having timeouts in the tests themselves is not always
desirable, as it may interfere with the debugger.  It is better to rely
on the test harness to stop long running tests.

Change-Id: I1ba1440028bd84d82222ff3650505562c0dd38c1